backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.9.0
snowflake-connector-python>=3.12.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
